+++ packages/cashc/src/semantic/TypeCheckTraversal.ts
@@ -259,16 +259,22 @@ export default class TypeCheckTraversal extends AstTraversal {
       case UnaryOperator.INPUT_OUTPOINT_INDEX:
       case UnaryOperator.INPUT_SEQUENCE_NUMBER:
       case UnaryOperator.OUTPUT_VALUE:
+      case UnaryOperator.INPUT_TOKEN_AMOUNT:
+      case UnaryOperator.OUTPUT_TOKEN_AMOUNT:
         expectInt(node, node.expression.type);
         node.type = PrimitiveType.INT;
         return node;
       case UnaryOperator.INPUT_LOCKING_BYTECODE:
       case UnaryOperator.INPUT_UNLOCKING_BYTECODE:
       case UnaryOperator.OUTPUT_LOCKING_BYTECODE:
+      case UnaryOperator.INPUT_NFT_COMMITMENT:
+      case UnaryOperator.OUTPUT_NFT_COMMITMENT:
         expectInt(node, node.expression.type);
         node.type = new BytesType();
         return node;
       case UnaryOperator.INPUT_OUTPOINT_HASH:
+      case UnaryOperator.INPUT_TOKEN_CATEGORY:
+      case UnaryOperator.OUTPUT_TOKEN_CATEGORY:
         expectInt(node, node.expression.type);
         node.type = new BytesType(32);
         return node;